Fix: delete API nil pointer issue

+++ database/point.go
@@ -271,14 +271,21 @@ func (d *GormDatabase) UpdatePointValue(pointModel *model.Point, priority *map[s
 }
 
 func (d *GormDatabase) DeletePoint(uuid string) (bool, error) {
-	point, _ := d.GetPoint(uuid, api.Args{})
+	point, err := d.GetPoint(uuid, api.Args{})
+	if err != nil {
+		return false, err
+	}
 	producers, _ := d.GetProducers(api.Args{ProducerThingUUID: &point.UUID})
-	for _, producer := range producers {
-		_, _ = d.DeleteProducer(producer.UUID)
+	if producers != nil {
+		for _, producer := range producers {
+			_, _ = d.DeleteProducer(producer.UUID)
+		}
 	}
 	writers, _ := d.GetWriters(api.Args{WriterThingUUID: &point.UUID})
-	for _, writer := range writers {
-		_, _ = d.DeleteWriter(writer.UUID)
+	if writers != nil {
+		for _, writer := range writers {
+			_, _ = d.DeleteWriter(writer.UUID)
+		}
 	}
 	query := d.DB.Delete(&point)
 	if query.Error != nil {
